README
The importance of education in society cannot be overstated, every government and country in the world strives to make this area better. The corona-virus outbreak, which disrupted life around the world in 2020, has had a variety of effects on educational systems; studies show that student performance has decreased since then, highlighting the need to take this issue more seriously and try to find effective solutions, as well as the influencing factors. Education is a critical component of society, and every government and country in the world strives to develop it. The corona-virus outbreak, which disrupted life around the world in 2020, has had a variety of effects on educational systems; studies show that student performance has decreased since then, highlighting the need to take this issue more seriously and try to find effective solutions, as well as the influencing factors. With the use of past student records, we were able to develop a model that would allow us to assist students in improving their exam performance by forecasting student success. So, obviously, it's a classification problem, and we'll classify a student based on the information he provides, as well as compare different classifiers such as KNN and SVM classifiers. Many factors influence a student's exam performance, such as family troubles or alcohol intake, and we want to use our machine learning talents to forecast whether or not a student will pass his final exam. And we developed the best classifier that is more accurate and avoids overfitting and underfitting by employing simple procedures. We also want to know which elements have the greatest impact on a student's success.